Move Footer to src/app/components
