+++ src/app/components/Footer/index.test.jsx
@@ -1,9 +1,6 @@
 import React from 'react';
-import Footer from '#components/Footer';
-import {
-  render,
-  screen,
-} from '../../../components/react-testing-library-with-providers';
+import Footer from '#app/components/Footer';
+import { render, screen } from '../react-testing-library-with-providers';
 
 describe('Footer', () => {
   const { trustProjectLink, externalLink, links, copyrightText } = {
